Expanding the filename and adding to the list correctly
Before this change, a path like "~/user/.emacs.d" caused problems with
LSP, because it did not like the tilde in the path. Now the path is
expanded.

This also fixes the call to `append`, which was previously creating a
single cons cell, where the `-javaagent` was the `cdr`. Now the `cdr`
is correctly a list, with this as a new argument.